backend/requirements.txt
Django==2.2.16
django-multiselectfield==0.1.12
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
django-filter==22.1
djangorestframework-jwt==1.11.0
Pillow==9.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
sqlparse==0.4.2
pytz==2022.1
PyJWT==1.7.1
asgiref==3.5.2
djoser==2.1.0
webcolors==1.12